backend/requirements.txt
flask
flask-cors
requests
lxml
beautifulsoup4
tldextract
readability-lxml